# internal/modules/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 九小二 Backend API
  version: 1.0.0-cp1
  description: 九小二产品一期 Go 后端唯一公开契约，覆盖交易、履约、打印、通知、账号运营与酒类合规。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/customer/send-code:
    post:
      summary: 发送 C 端 mock 验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendCodeReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/customer/sms-login:
    post:
      summary: C 端验证码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SmsLoginReq" }
      responses:
        "200": { $ref: "#/components/responses/Token" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/customer/wechat-login:
    post:
      summary: 微信小程序登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WeChatLoginReq" }
      responses:
        "200": { $ref: "#/components/responses/Token" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/customer/phone-bind:
    post:
      summary: 绑定微信授权手机号
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PhoneBindReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/admin/login:
    post:
      summary: 管理员登录
      security: []
      requestBody: { $ref: "#/components/requestBodies/PasswordLogin" }
      responses:
        "200": { $ref: "#/components/responses/Token" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/merchant/login:
    post:
      summary: 商户登录
      security: []
      requestBody: { $ref: "#/components/requestBodies/PasswordLogin" }
      responses:
        "200": { $ref: "#/components/responses/Token" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/rider/send-code:
    post:
      summary: 发送骑手登录或申请用的 mock 验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendCodeReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/rider/sms-login:
    post:
      summary: 已审核骑手使用手机号和短信验证码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SmsLoginReq" }
      responses:
        "200": { $ref: "#/components/responses/Token" }
        default: { $ref: "#/components/responses/Problem" }
...
  /auth/refresh:
    post:
      summary: 刷新 token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshReq" }
      responses:
        "200": { $ref: "#/components/responses/Token" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      summary: 登出并写入 JWT 黑名单
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
...
  /orders:
    get:
      summary: C 端本人订单列表
      x-permission: order:list
      x-object-scope: customer.self_orders
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - { name: status, in: query, schema: { type: string, maxLength: 32 } }
        - { name: order_by, in: query, schema: { type: string, enum: ["created_at desc,id desc"] } }
      responses:
        "200":
          description: 本人订单分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerOrderPageEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      summary: 创建订单并预占库存
      x-permission: order:create
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCreateReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}:
    get:
      summary: 订单详情
      x-permission: order:view
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
...
  /orders/{id}/verification:
    get:
      summary: 顾客受控查看送达码
      x-permission: delivery_verification:view_customer
      x-object-scope: customer_self_delivering_order
      x-sensitive-response: true
      parameters: [ { $ref: "#/components/parameters/ID" } ]
      responses:
        "200":
          description: 送达码敏感响应，禁止缓存
          headers:
            Cache-Control:
              description: 必须为 no-store
              x-required-response-header: true
              schema: { type: string, enum: [no-store] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeliveryVerificationEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: application_access JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
    PageToken:
      name: page_token
      in: query
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
...
    Token:
      description: 登录或刷新成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TokenEnvelope" }
...
    Problem:
      description: 标准业务或系统错误
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  requestBodies:
...
    PasswordLogin:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PasswordLoginReq" }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { type: object, additionalProperties: true, nullable: true }
        request_id: { type: string }
...
    TokenEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          required: [access_token, refresh_token, expires_in, account_type, account_id, profile]
          properties:
            access_token: { type: string }
            refresh_token: { type: string }
            expires_in: { type: integer, format: int64 }
            account_type: { type: string, enum: [customer, admin, merchant, rider] }
            account_id: { type: string }
            profile: { type: object, additionalProperties: true }
        request_id: { type: string }
...
    SendCodeReq:
      type: object
      required: [phone]
      properties:
        phone: { type: string }
    SmsLoginReq:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string }
        code: { type: string }
    WeChatLoginReq:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 4, maxLength: 256 }
        device_id: { type: string, maxLength: 128 }
    PhoneBindReq:
      type: object
      required: [phone_code]
      properties:
        phone_code: { type: string, minLength: 4, maxLength: 256 }
    PasswordLoginReq:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    OrderCreateReq:
      type: object
      required: [shop_id, address_id, items]
      properties:
        shop_id: { type: string }
        address_id: { type: string }
        items:
          type: array
          items:
            type: object
            required: [shop_product_id, quantity]
            properties:
              shop_product_id: { type: string }
              quantity: { type: integer }
    ShopSummary:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        name: { type: string }
    OrderItemSummary:
      type: object
      additionalProperties: false
      required: [product_id, name, quantity]
      properties:
        product_id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        name: { type: string }
        spec: { type: string }
        image_url: { type: string }
        quantity: { type: integer, minimum: 0 }
    OrderSummary:
      type: object
      additionalProperties: false
      required: [id, order_no, shop_id, status, pay_status, delivery_status, payable_amount, shop_summary, item_summary, item_kind_count, total_quantity, created_at, updated_at]
      properties:
        id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        order_no: { type: string }
        shop_id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        status: { type: string }
        pay_status: { type: string }
        delivery_status: { type: string }
        payable_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        shop_summary: { $ref: "#/components/schemas/ShopSummary" }
        item_summary: { $ref: "#/components/schemas/OrderItemSummary" }
        item_kind_count: { type: integer, minimum: 0 }
        total_quantity: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CustomerOrderPageEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data:
          type: object
          additionalProperties: false
          required: [items, next_page_token]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/OrderSummary" }
            next_page_token: { type: string }
        request_id: { type: string }
...
    RefreshReq:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
...
    DeliveryVerification:
      type: object
      additionalProperties: false
      required: [delivery_order_id, stage, status, code, code_mask, expires_at, version]
      properties:
        delivery_order_id: { type: string }
        stage: { type: string, enum: [delivery] }
        status: { type: string, enum: [active, locked] }
        code: { type: string, pattern: "^[0-9]{6}$", writeOnly: false, x-sensitive: one_time_credential }
        code_mask: { type: string, pattern: "^\\*{4}[0-9]{2}$" }
        failed_attempts: { type: integer, minimum: 0 }
        remaining_attempts: { type: integer, minimum: 0 }
        expires_at: { type: string, format: date-time }
        locked_until: { type: string, format: date-time }
        verified_at: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
    DeliveryVerificationEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data: { $ref: "#/components/schemas/DeliveryVerification" }
        request_id: { type: string }
...
    ProblemDetails:
      type: object
      required: [type, title, status, error_code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error_code:
          type: string
          description: 幂等写请求同 key 异参固定返回 IDEMPOTENCY_KEY_REUSED；同参请求仍处于有效处理租约时固定返回 IDEMPOTENCY_IN_PROGRESS。
        request_id: { type: string }